+++ CHEHBI_TO_DO_LIST/src/main/java/com/example/CHEHBI_TO_DO_LIST/services/TodoItemService.java
@@ -7,8 +7,8 @@ public interface TodoItemService {
     Iterable<TodoItem> getAllTasks();
 
     void saveTask(TodoItem todoItem);
-    TodoItem findTodoItemById(Long id);
+    TodoItem findTodoItemById(int id);
     void deleteTask(TodoItem todoItem);
-    void updateCompleteStatus(Long itemId, boolean isComplete);
+    void updateCompleteStatus(int itemId, boolean isComplete);
 
 }
